requirements.txt
accelerate==0.30.1
decord==0.6.0
deepspeed==0.9.5
Jinja2==3.1.4
ninja==1.11.1.1
numpy==1.26.4
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.535.161
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
pillow==10.4.0
torch==2.3.1
torchaudio==2.3.1
torchvision
tqdm==4.66.4
transformers==4.41.1
xformers
